requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
celery==4.4.2
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
kombu==4.6.11
packaging==21.3
prompt-toolkit==3.0.24
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
vine==1.3.0
wcwidth==0.2.5
wrapt==1.13.3
